+++ src/actions/index.js
@@ -1,10 +1,11 @@
-import bindAction from './bind';
+import bindAction, { bindSpecificComponentsAction } from './bind';
 import fetchAction from './fetch';
 import importAction from './import';
 import watchAction from './watch';
 
 export {
   bindAction,
+  bindSpecificComponentsAction,
   fetchAction,
   importAction,
   watchAction,