+++ src/components-map/index.js
@@ -1,7 +1,8 @@
-import { buildForInline, build, buildForNamespaces } from './components-map';
+import { buildForInline, build, buildForNamespaces, generateId } from './components-map';
 
 export {
   buildForInline,
   build,
   buildForNamespaces,
+  generateId,
 };